Adjust input values to time spent for updating

+++ FractalPlotter/Graphics/Camera.cs
@@ -85,7 +85,7 @@ namespace FractalPlotter.Graphics
 		/// Move in the specified directions. The values will be multiplied by the camera speed as well.
 		/// In order to allow movement in deeper regions, moving is also being divided by the current scale.
 		/// </summary>
-		public static void Translate(int x, int y, int z)
+		public static void Translate(float x, float y, float z)
 		{
 			var speed = RelativeSpeed;
 			Location += new Vector3(x * speed, y * speed, z * speed);
